+++ app/views/calendar/_event_cell.html.erb
@@ -5,9 +5,17 @@
           - <%= current_event.end_time.strftime('%H:%M') %></div>
         <div class="event-name"><%= current_event.name %></div>
         <div class="event-user">
-          <small>by</small> <%= current_event.user %></div>
+          <% if current_event.user %>
+          <small>by</small> <%= current_event.user %>
+          <% end %>
+        </div>
         <div class="event-location">
-          <small>in</small> <%= current_event.conference_room.title %></div>
+          <% if current_event.conference_room&.title %>
+            <small>in</small> <%= current_event.conference_room.title %>
+          <% else %>
+              Occupied
+          <% end %>
+        </div>
       </div>
     </td>
 <% elsif !event_ongoing?(day, time) %>
