+++ spec/helpers/events_helper_spec.rb
@@ -17,19 +17,19 @@ RSpec.describe EventsHelper, type: :helper do
 
     context 'nil given' do
       it 'returns day from the following week' do
-        expect(helper.next_week(nil)).to eq(Date.today + 7.days)
+        expect(helper.next_week(nil)).to eq(Date.today.next_week)
       end
     end
 
     context 'date given in string format' do
       it 'returns date 7 days after given day' do
-        expect(helper.next_week(custom_day_s)).to eq(custom_day + 7.days)
+        expect(helper.next_week(custom_day_s)).to eq(custom_day.next_week)
       end
     end
 
     context 'no string given' do
       it 'returns day from the following week' do
-        expect(helper.next_week(custom_day)).to eq(Date.today + 7.days)
+        expect(helper.next_week(custom_day)).to eq(custom_day.next_week)
       end
     end
   end
@@ -40,19 +40,19 @@ RSpec.describe EventsHelper, type: :helper do
 
     context 'nil given' do
       it 'returns day from the previous week' do
-        expect(helper.previous_week(nil)).to eq(Date.today - 7.days)
+        expect(helper.previous_week(nil)).to eq(Date.today.prev_week)
       end
     end
 
     context 'date given in string format' do
       it 'returns date 7 days before given day' do
-        expect(helper.previous_week(custom_day_s)).to eq(custom_day - 7.days)
+        expect(helper.previous_week(custom_day_s)).to eq(custom_day.prev_week)
       end
     end
 
     context 'no string given' do
       it 'returns day from the previous week' do
-        expect(helper.previous_week(custom_day)).to eq(Date.today - 7.days)
+        expect(helper.previous_week(custom_day)).to eq(custom_day.prev_week)
       end
     end
   end
